Convert to mambaforge
The full installation is now based on mambaforge instead of Anaconda, but Anaconda packages are installed later via mamba.

Now uses a JupyterHub config file that only includes the necessary pieces. The boilerplate and explanation is removed.

The details to setup nbgrader are included. The old details used to configure Azure AD added in comments.